Added sprites and tools folder

+++ MAKERphone.h
@@ -278,7 +278,7 @@ public:
   uint8_t backgroundIndex = 0;
   uint8_t volume = 10; //volume 0-14
   bool pinLock = 0;
-  uint16_t pinNumber = 1234;
+  uint8_t pinNumber = 1234;
   bool simInserted = 0;
   uint32_t sleepTimer = millis();
